query_rewrite/requirements.txt
# Evolved RuleSelector — Dependencies
# Outer loop (evolve_loop/) only. R-Bot inner loop has separate requirements.

anthropic>=0.39.0
jpype1>=1.5.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
